chore: reorganise and cleanup
