+++ packages/chain-adapters/src/api.ts
@@ -18,11 +18,6 @@ import type {
   ValidAddressResult,
 } from './types'
 
-/**
- * Type alias for a Map that can be used to manage instances of ChainAdapters
- */
-export type ChainAdapterManager = Map<ChainId, ChainAdapter<ChainId>>
-
 export type ChainAdapter<T extends ChainId> = {
   /**
    * Internal use only. The chain-adapter name for e.g logging purposes
