--- src/components/EarnDashboard/components/ProviderDetails/LpPositionsByAsset.tsx
@@ -1,146 +1,0 @@
-import { Button, Flex } from '@chakra-ui/react'
-import { Tag } from '@chakra-ui/tag'
-import { fromAssetId } from '@shapeshiftoss/caip'
-import { DefiAction } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
-import qs from 'qs'
-import { useCallback, useMemo } from 'react'
-import { useTranslate } from 'react-polyglot'
-import { useHistory, useLocation } from 'react-router'
-import type { Column, Row } from 'react-table'
-import { Amount } from 'components/Amount/Amount'
-import { ReactTable } from 'components/ReactTable/ReactTable'
-import { AssetCell } from 'components/StakingVaults/Cells'
-import { WalletActions } from 'context/WalletProvider/actions'
-import { useWallet } from 'hooks/useWallet/useWallet'
-import { trackOpportunityEvent } from 'lib/mixpanel/helpers'
-import { MixPanelEvents } from 'lib/mixpanel/types'
-import type { LpEarnOpportunityType, OpportunityId } from 'state/slices/opportunitiesSlice/types'
-import { selectAggregatedEarnUserLpOpportunities, selectAssets } from 'state/slices/selectors'
-import { useAppSelector } from 'state/store'
-
-type LpPositionsByAssetProps = {
-  ids: OpportunityId[]
-}
-
-export type RowProps = Row<LpEarnOpportunityType>
-
-export const LpPositionsByAsset: React.FC<LpPositionsByAssetProps> = ({ ids }) => {
-  const translate = useTranslate()
-  const location = useLocation()
-  const history = useHistory()
-  const {
-    state: { isConnected, isDemoWallet },
-    dispatch,
-  } = useWallet()
-  const assets = useAppSelector(selectAssets)
-  const lpOpportunities = useAppSelector(selectAggregatedEarnUserLpOpportunities)
-  const filteredDown = lpOpportunities.filter(e => ids.includes(e.assetId as OpportunityId))
-
-  const handleClick = useCallback(
-    (row: RowProps, action: DefiAction) => {
-      const { original: opportunity } = row
-      const {
-        type,
-        provider,
-        contractAddress,
-        chainId,
-        rewardAddress,
-        assetId,
-        highestBalanceAccountAddress,
-      } = opportunity
-      const { assetReference, assetNamespace } = fromAssetId(assetId)
-
-      if (!isConnected && isDemoWallet) {
-        dispatch({ type: WalletActions.SET_WALLET_MODAL, payload: true })
-        return
-      }
-
-      trackOpportunityEvent(
-        MixPanelEvents.ClickOpportunity,
-        {
-          opportunity,
-          element: 'Table Row',
-        },
-        assets,
-      )
-
-      history.push({
-        pathname: location.pathname,
-        search: qs.stringify({
-          type,
-          provider,
-          chainId,
-          contractAddress,
-          assetNamespace,
-          assetReference,
-          highestBalanceAccountAddress,
-          rewardId: rewardAddress,
-          modal: action ? action : 'overview',
-        }),
-        state: { background: location },
-      })
-    },
-    [assets, dispatch, history, isConnected, isDemoWallet, location],
-  )
-  const columns: Column<LpEarnOpportunityType>[] = useMemo(
-    () => [
-      {
-        Header: translate('defi.liquidityPool'),
-        accessor: 'assetId',
-        Cell: ({ row }: { row: RowProps }) => (
-          <AssetCell
-            assetId={row.original.underlyingAssetId ?? row.original.assetId}
-            subText={row.original.version}
-            icons={row.original.icons}
-            opportunityName={row.original.opportunityName}
-            showAssetSymbol={row.original.showAssetSymbol}
-            isExternal={row.original.isReadOnly}
-          />
-        ),
-        disableSortBy: true,
-      },
-      {
-        Header: translate('defi.totalValue'),
-        accessor: 'fiatAmount',
-        Cell: ({ row }: { row: RowProps }) => <Amount.Fiat value={row.original.fiatAmount} />,
-      },
-      {
-        Header: translate('defi.apy'),
-        accessor: 'apy',
-        Cell: ({ row }: { row: RowProps }) => (
-          <Tag colorScheme='green'>
-            <Amount.Percent value={row.original.apy} />
-          </Tag>
-        ),
-      },
-      {
-        Header: () => null,
-        id: 'expander',
-        Cell: ({ row }: { row: RowProps }) => (
-          <Flex gap={4} justifyContent='flex-end' width='full'>
-            <Button
-              variant='ghost'
-              width={{ base: 'full', md: 'auto' }}
-              size='sm'
-              colorScheme='blue'
-              onClick={() => handleClick(row, DefiAction.Overview)}
-            >
-              {translate('common.manage')}
-            </Button>
-          </Flex>
-        ),
-      },
-    ],
-    [handleClick, translate],
-  )
-
-  if (!filteredDown.length) return null
-
-  return (
-    <ReactTable
-      data={filteredDown}
-      columns={columns}
-      initialState={{ sortBy: [{ id: 'fiatAmount', desc: true }] }}
-    />
-  )
-}
