+++ src/components/MultiHopTrade/components/TradeConfirm/TradeConfirm.tsx
@@ -22,13 +22,13 @@ import { useHistory } from 'react-router-dom'
 import { Amount } from 'components/Amount/Amount'
 import { Card } from 'components/Card/Card'
 import { HelperTooltip } from 'components/HelperTooltip/HelperTooltip'
+import { AssetToAsset } from 'components/MultiHopTrade/components/TradeConfirm/AssetToAsset'
+import { ReceiveSummary } from 'components/MultiHopTrade/components/TradeConfirm/ReceiveSummary'
 import { getMixpanelEventData } from 'components/MultiHopTrade/helpers'
 import { useTradeExecution } from 'components/MultiHopTrade/hooks/useTradeExecution/useTradeExecution'
 import { Row } from 'components/Row/Row'
 import { SlideTransition } from 'components/SlideTransition'
 import { RawText, Text } from 'components/Text'
-import { AssetToAsset } from 'components/Trade/TradeConfirm/AssetToAsset'
-import { ReceiveSummary } from 'components/Trade/TradeConfirm/ReceiveSummary'
 import { chainSupportsTxHistory } from 'components/Trade/utils'
 import { WithBackButton } from 'components/Trade/WithBackButton'
 import { WalletActions } from 'context/WalletProvider/actions'
