+++ src/components/MultiHopTrade/components/TradeInput/TradeInput.tsx
@@ -17,6 +17,7 @@ import { useFormContext } from 'react-hook-form'
 import { useTranslate } from 'react-polyglot'
 import { useHistory } from 'react-router'
 import { MessageOverlay } from 'components/MessageOverlay/MessageOverlay'
+import { ReceiveSummary } from 'components/MultiHopTrade/components/TradeConfirm/ReceiveSummary'
 import { DonationCheckbox } from 'components/MultiHopTrade/components/TradeInput/components/DonationCheckbox'
 import { ManualAddressEntry } from 'components/MultiHopTrade/components/TradeInput/components/ManualAddressEntry'
 import { getMixpanelEventData } from 'components/MultiHopTrade/helpers'
@@ -29,7 +30,6 @@ import { Text } from 'components/Text'
 import { TradeAssetSelect } from 'components/Trade/Components/AssetSelection'
 import { RateGasRow } from 'components/Trade/Components/RateGasRow'
 import { TradeAssetInput } from 'components/Trade/Components/TradeAssetInput'
-import { ReceiveSummary } from 'components/Trade/TradeConfirm/ReceiveSummary'
 import { useErrorHandler } from 'hooks/useErrorToast/useErrorToast'
 import { useModal } from 'hooks/useModal/useModal'
 import { useToggle } from 'hooks/useToggle/useToggle'
