+++ src/components/MultiHopTrade/hooks/useGetTradeQuotes/typeGuards.ts
