--- src/components/Trade/hooks/useTradeRoutes/types.ts
@@ -1,4 +1,0 @@
-export enum AssetClickAction {
-  Buy = 'buy',
-  Sell = 'sell',
-}
