+++ src/state/slices/opportunitiesSlice/types.ts
@@ -164,8 +164,6 @@ export type GetOpportunityMetadataOutput = {
   type: DefiType
 }
 
-export type GetReadOnlyOpportunitiesOutput = ReadOnlyOpportunityType[] | undefined
-
 export type GetOpportunityUserDataOutput = {
   byAccountId: OpportunitiesState[DefiType]['byAccountId']
   type: DefiType
@@ -238,13 +236,6 @@ export type AggregatedOpportunitiesByAssetIdReturn = {
   cryptoBalancePrecision: string
   fiatRewardsAmount: string
   opportunities: Record<DefiType, OpportunityId[]>
-}
-
-export type AggregatedOpportunitiesByFilterReturn = {
-  id: OpportunityId
-  type: string
-  fiatAmount: string
-  cryptoBalancePrecision: string
 }
 
 export type AggregatedOpportunitiesByProviderReturn = {
